requirements.txt
baidu-aip
numpy